test/acceptance_with_python/requirements.txt
git+https://github.com/weaviate/weaviate-python-client.git@b2f2e360e61de539e98a61c60a1963f78a5fa1ee

pytest>=8.0.1,<9.0.0
pytest-xdist==3.6.1
pytest-benchmark==4.0.0

black>=24.2.0,<25.0.0
loguru>=0.7.2

numpy==2.1.2
pyyaml>=6.0.1,<7.0.0
